+++ src/main/java/com/onlinecode/itnhaque/repository/RoleRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import com.onlinecode.itnhaque.domain.Role;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long>, JpaSpecificationExecutor<Role> {
+public interface RoleRepository extends JpaRepository<Role, Integer>, JpaSpecificationExecutor<Role> {
     boolean existsByName(String name);
 
     Role findByName(String name);
